+++ lib/cocoapods/config.rb
@@ -17,11 +17,6 @@ module Pod
       :skip_repo_update    => false,
       :skip_download_cache => !ENV['COCOAPODS_SKIP_CACHE'].nil?,
 
-      :clean               => true,
-      :integrate_targets   => true,
-      :deduplicate_targets => true,
-      :deterministic_uuids => ENV['COCOAPODS_DISABLE_DETERMINISTIC_UUIDS'].nil?,
-      :lock_pod_source     => true,
       :new_version_message => ENV['COCOAPODS_SKIP_UPDATE_MESSAGE'].nil?,
 
       :cache_root          => Pathname.new(Dir.home) + 'Library/Caches/CocoaPods',
@@ -73,34 +68,6 @@ module Pod
     #-------------------------------------------------------------------------#
 
     # @!group Installation
-
-    # @return [Bool] Whether the installer should clean after the installation.
-    #
-    attr_accessor :clean
-    alias_method :clean?, :clean
-
-    # @return [Bool] Whether the installer should remove write permissions for
-    #                installed pod source files after the installation.
-    #
-    attr_accessor :lock_pod_source
-    alias_method :lock_pod_source?, :lock_pod_source
-
-    # @return [Bool] Whether CocoaPods should integrate a user target and build
-    #         the workspace or just create the Pods project.
-    #
-    attr_accessor :integrate_targets
-    alias_method :integrate_targets?, :integrate_targets
-
-    # @return [Bool] Whether CocoaPods should deduplicate pod targets.
-    #
-    attr_accessor :deduplicate_targets
-    alias_method :deduplicate_targets?, :deduplicate_targets
-
-    # @return [Bool] Whether CocoaPods should give the pods project
-    #                deterministic UUIDs.
-    #
-    attr_accessor :deterministic_uuids
-    alias_method :deterministic_uuids?, :deterministic_uuids
 
     # @return [Bool] Whether the installer should skip the repos update.
     #