+++ lib/cocoapods/installer/podfile_validator.rb
@@ -12,6 +12,10 @@ module Pod
       #
       attr_reader :errors
 
+      # @return [Array<String>] any warnings that have occured during the validation
+      #
+      attr_reader :warnings
+
       # Initialize a new instance
       # @param [Podfile] podfile
       #        The podfile to validate
@@ -19,6 +23,7 @@ module Pod
       def initialize(podfile)
         @podfile = podfile
         @errors = []
+        @warnings = []
         @validated = false
       end
 
@@ -27,6 +32,8 @@ module Pod
       #
       def validate
         validate_pod_directives
+        validate_no_abstract_only_pods!
+        validate_dependencies_are_present!
 
         @validated = true
       end
@@ -38,7 +45,7 @@ module Pod
       def valid?
         validate unless @validated
 
-        @validated && errors.size == 0
+        @validated && errors.empty?
       end
 
       # A message describing any errors in the
@@ -52,6 +59,10 @@ module Pod
 
       def add_error(error)
         errors << error
+      end
+
+      def add_warning(warning)
+        warnings << warning
       end
 
       def validate_pod_directives
@@ -79,6 +90,36 @@ module Pod
         if pod_spec_or_path && specified_downloaders.size > 0
           add_error "The dependency `#{dependency.name}` specifies `podspec` or `path` in combination with other" \
             ' download strategies. This is not allowed'
+        end
+      end
+
+      # Warns the user if the podfile is empty.
+      #
+      # @note   The workspace is created in any case and all the user projects
+      #         are added to it, however the projects are not integrated as
+      #         there is no way to discern between target definitions which are
+      #         empty and target definitions which just serve the purpose to
+      #         wrap other ones. This is not an issue because empty target
+      #         definitions generate empty libraries.
+      #
+      # @return [void]
+      #
+      def validate_dependencies_are_present!
+        if podfile.target_definitions.values.all?(&:empty?)
+          add_warning 'The Podfile does not contain any dependencies.'
+        end
+      end
+
+      # Verifies that no dependencies in the Podfile will end up not being built
+      # at all. In other words, all dependencies _must_ belong to a non-abstract
+      # target, or be inherited by a target where `inheritance == complete`.
+      #
+      def validate_no_abstract_only_pods!
+        all_dependencies = podfile.dependencies
+        concrete_dependencies = podfile.target_definition_list.reject(&:abstract?).flat_map(&:dependencies).uniq
+        abstract_only_dependencies = all_dependencies - concrete_dependencies
+        abstract_only_dependencies.each do |dep|
+          add_error "The dependency `#{dep}` is not used in any concrete target."
         end
       end
     end