+++ spec/integration.rb
@@ -44,9 +44,12 @@ $:.unshift((ROOT + 'spec').to_s)
 
 require 'rubygems'
 require 'bundler/setup'
+
 require 'pretty_bacon'
 require 'colored'
 require 'clintegracon'
+
+require 'cocoapods-core/yaml_helper'
 require 'fileutils'
 require 'integration/file_tree'
 require 'integration/xcodeproj_project_yaml'
@@ -78,11 +81,19 @@ CLIntegracon.configure do |c|
   end
 
   # Register special handling for YAML files
-  c.transform_produced '**/{Podfile,Manifest}.lock' do |path|
-    # Remove CocoaPods version
+  c.transform_produced %r{(^|/)(Podfile|Manifest).lock$} do |path|
+    # Remove CocoaPods version & Podfile checksum
     yaml = YAML.load(path.read)
     yaml.delete('COCOAPODS')
-    path.open('w') { |f| f << YAML.dump(yaml) }
+    yaml.delete('PODFILE CHECKSUM')
+    keys_hint = [
+      'PODS',
+      'DEPENDENCIES',
+      'EXTERNAL SOURCES',
+      'CHECKOUT OPTIONS',
+      'SPEC CHECKSUMS',
+    ]
+    path.open('w') { |f| f << Pod::YAMLHelper.convert_hash(yaml, keys_hint, "\n\n") }
   end
 
   # So we don't need to compare them directly
@@ -150,6 +161,22 @@ describe_cli 'pod' do
   describe 'Pod install' do
     # Test installation with no integration
     # Test subspecs inheritance
+
+    #--------------------------------------#
+
+    describe 'Pod init' do
+      describe 'Initializes a Podfile with a single platform' do
+        behaves_like cli_spec 'init_single_platform',
+                              'init'
+      end
+    end
+
+    #--------------------------------------#
+
+    describe 'Integrates a project with an empty Podfile with CocoaPods' do
+      behaves_like cli_spec 'install_no_dependencies',
+                            'install --no-repo-update'
+    end
 
     describe 'Integrates a project with CocoaPods' do
       behaves_like cli_spec 'install_new',
@@ -294,15 +321,6 @@ describe_cli 'pod' do
     describe 'Lints a Pod' do
       behaves_like cli_spec 'spec_lint',
                             'spec lint --quick'
-    end
-  end
-
-  #--------------------------------------#
-
-  describe 'Pod init' do
-    describe 'Initializes a Podfile with a single platform' do
-      behaves_like cli_spec 'init_single_platform',
-                            'init'
     end
   end