+++ spec/unit/installer/user_project_integrator/target_integrator/xcconfig_integrator_spec.rb
@@ -8,7 +8,6 @@ module Pod
       Project.new(config.sandbox.project_path).save
       @target = @project.targets.first
       target_definition = Podfile::TargetDefinition.new('Pods', nil)
-      target_definition.link_with_first_target = true
       @pod_bundle = AggregateTarget.new(target_definition, config.sandbox)
       @pod_bundle.user_project = @project
       @pod_bundle.client_root = project_path.dirname
@@ -108,6 +107,18 @@ module Pod
 
     it 'handles when xcconfig is set to another sandbox xcconfig' do
       group = @project.new_group('Pods')
+
+      old_config = group.new_file('../Pods/Target Support Files/Pods-Foo/SampleConfig.xcconfig')
+      @target.build_configurations.each do |config|
+        config.base_configuration_reference = old_config
+      end
+      XCConfigIntegrator.integrate(@pod_bundle, [@target])
+      @target.build_configurations.each do |config|
+        config.base_configuration_reference.should.not == old_config
+        config.base_configuration_reference.path.should == @pod_bundle.xcconfig_relative_path(config.name)
+      end
+
+      @pod_bundle.stubs(:label).returns('Pods-Foo')
       old_config = group.new_file('../Pods/Target Support Files/Pods/SampleConfig.xcconfig')
       @target.build_configurations.each do |config|
         config.base_configuration_reference = old_config