+++ spec/unit/library_spec.rb
@@ -5,7 +5,6 @@ module Pod
     describe 'In general' do
       before do
         @target_definition = Podfile::TargetDefinition.new('Pods', nil)
-        @target_definition.link_with_first_target = true
         @lib = AggregateTarget.new(@target_definition, config.sandbox)
       end
 
@@ -29,7 +28,6 @@ module Pod
     describe 'Support files' do
       before do
         @target_definition = Podfile::TargetDefinition.new('Pods', nil)
-        @target_definition.link_with_first_target = true
         @lib = AggregateTarget.new(@target_definition, config.sandbox)
         @lib.client_root = config.sandbox.root.dirname
       end